tutorials/aws-kinesis-dynamodb/sink_dynamodb/requirements.txt
python-dotenv==1.0.1
boto3==1.34.83
Faker==24.8.0
glassflow